third_party/blink/web_tests/platform/linux-chrome/external/wpt/reporting/cross-origin-reports-isolated.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Reports were sent in two requests.
  assert_equals: expected 2 but got 0
Harness: the test ran to completion.
